requirements.txt
django==2.2.7
djangorestframework==3.10.3
gunicorn==20.0.3
pytest-django==3.7.0
Faker==2.0.4
assertpy==0.15
django-cors-headers==3.2.0
requests==2.22.0
